Fix: 관리자용 계곡 정보 수정 페이지 오류 수정

+++ tovalley-server/src/main/java/kr/ac/kumoh/illdang100/tovalley/domain/water_place/WaterPlaceDetail.java
@@ -53,7 +53,7 @@ public class WaterPlaceDetail {
     private Double turbidity; // 탁도(NTU)
 
     public void update(WaterPlaceForm form) {
-        Double waterPlaceSegment = form.getWaterPlaceSegment();
+        Integer waterPlaceSegment = form.getWaterPlaceSegment();
         Double annualVisitors = form.getAnnualVisitors();
         Integer dangerSegments = form.getDangerSegments();
         Integer dangerSignboardsNum = form.getDangerSignboardsNum();
@@ -73,7 +73,7 @@ public class WaterPlaceDetail {
 
     public static WaterPlaceDetail createNewWaterPlaceDetail(WaterPlace waterPlace, CreateWaterPlaceForm form) {
 
-        Double waterPlaceSegment = form.getWaterPlaceSegment();
+        Integer waterPlaceSegment = form.getWaterPlaceSegment();
         Double avgDepth = form.getAvgDepth();
         Double annualVisitors = form.getAnnualVisitors();
         Integer dangerSegments = form.getDangerSegments();
